feat: migrates repo to typescript, adds linting and solhint (#9)

+++ interfaces/keep3r/IKeep3rV1Helper.sol
@@ -2,5 +2,5 @@
 pragma solidity 0.6.12;
 
 interface IKeep3rV1Helper {
-    function getQuoteLimit(uint gasUsed) external view returns (uint);
+    function getQuoteLimit(uint256 gasUsed) external view returns (uint256);
 }
