+++ interfaces/keep3r/IUniswapV2SlidingOracle.sol
@@ -2,8 +2,15 @@
 pragma solidity 0.6.12;
 
 interface IUniswapV2SlidingOracle {
-    function current(address tokenIn, uint amountIn, address tokenOut) external view returns (uint);
+    function current(
+        address tokenIn,
+        uint256 amountIn,
+        address tokenOut
+    ) external view returns (uint256);
+
     function updatePair(address pair) external returns (bool);
+
     function workable(address pair) external view returns (bool);
+
     function workForFree() external;
 }
